+++ src/components/App/App.js
@@ -7,7 +7,7 @@ import {
   from,
 } from '@apollo/client';
 import { onError } from '@apollo/client/link/error';
-import Home from './Pages/Home';
+import Home from '../Home/Home';
 
 const errorLink = onError(({ graphqlErrors, networkError }) => {
   if (graphqlErrors) {
